feat: add base62 hashing function helper for unique link generation

+++ database/migrations/2026_08_28_121817_create_clicks_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('clicks', function (Blueprint $table) {
             $table->id();
             $table->string('url_id');
-            $table->integer('redirect_time_ms')->default(0);
+            $table->float('redirect_time_ms', 2)->default(0.00);
             $table->timestamps();
         });
     }
